Refactor UI's to avoid silencing some messages
Some failures happen in the test suite after this change, specially
since some "Resolving dependencies..." messages appear now in the
screen. To solve those, instead of asserting the "Resolving
dependencies..." message, we make sure we actually resolve before
running `bundle exec`. Also, we add a spec to make sure that the
autoresolve feature is signaled with a screen notice.